+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Coming Up
 
 - Auto-running the swing when a secondary file is saved
+- Grouping tutorial steps into sections (https://github.com/sveltejs/svelte/tree/master/site/content/examples)
+
+# v0.0.16
+
+- Added support for React Native web 🚀
+- Markup files/components can now be named `App.<extension>` in addition to `index.<extension>`
 
 # v0.0.15 (02/27/2021)